requirements.txt
gunicorn==19.9.0
itsdangerous==1.1.0
MarkupSafe==1.1.1
Werkzeug==0.15.5
tensorflow-cpu==2.3.0
Pillow>=5.4.0
numpy==1.17.2
Flask==1.1.2
h5py==2.10.0
emoji==1.2.0